requirements.txt
Package             Version
------------------- -------
asgiref             3.4.1
autopep8            1.5.7
Django              3.2.7
djangorestframework 3.12.4
pip                 21.1.1
pycodestyle         2.7.0
pytz                2021.1
setuptools          56.0.0
sqlparse            0.4.2
toml                0.10.2
